fix: keep the Remotion graph out of the transition vocabulary
Deriving Metadata from a schema made lib/project/types.ts import
TRANSITION_TYPES as a value rather than a type. Type imports are erased;
value imports are not, so the API routes that import voiceSearchParamsSchema
from that module started pulling @remotion/transitions into the server
bundle, and `next build` failed collecting page data for /api/v1/tts/voices.

Splits the Remotion presentation bindings into transitionPresentations.ts
and leaves transitions.ts as pure data, mirroring aspectRatio.ts. Naming a
transition no longer reaches a renderer. getPresentation had a single
consumer, so VideoComposition.tsx is the only import that moves.
